Colorguard — Photoshop CC
Brand new features!

1. Target a single PSD, or a folder of PSDs. If you close all of your
PSDs and run the script, you will be asked to select a folder — then the
script will update (but not save) all of the PSDs inside of it. If you
have a PSD open, the script will only target that PSD.

2. The script now targets text layers too! It will replace colors in
your text just like your vector shapes.

3. Optimized a bunch of code so it's even faster.